Renamed `/engine/metadata.go -> types.go` and moved `ExecutionError` into it
